requirements.txt
# Requirements for Student Behavior Analysis System
# Install with: pip install -r requirements.txt

# Core ML libraries - PyTorch
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                # PyTorch Image Models (for Swin Transformer)

# Core ML libraries - TensorFlow (legacy support)
tensorflow>=2.10.0
tensorflow-addons>=0.20.0
keras>=2.10.0

# Computer Vision
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
scikit-image>=0.19.0

# YOLO for object/person detection
ultralytics>=8.0.0

# InsightFace for facial recognition (Buffalo_l model)
insightface>=0.7.3
onnxruntime>=1.15.0

# Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pillow>=9.0.0

# Utilities
tqdm>=4.62.0
jupyter>=1.0.0
ipython>=7.0.0

# Web Framework
Flask>=2.3.0
flask-cors>=4.0.0
Werkzeug>=2.3.0

# Database - Hybrid AWS DynamoDB + MongoDB
boto3>=1.28.0              # AWS SDK for DynamoDB
pymongo>=4.5.0             # MongoDB driver
python-dotenv>=1.0.0       # Environment variables

# Optional: For Kaggle dataset download (if needed)
# kagglehub>=0.1.0
